+++ beaker/data_model/experiment.py
@@ -24,6 +24,10 @@ class Experiment(BaseModel):
     def display_name(self) -> str:
         return self.name if self.name is not None else self.id
 
+    @property
+    def workspace(self) -> WorkspaceRef:
+        return self.workspace_ref
+
 
 class Task(BaseModel):
     id: str